+++ src/simulator-mpi/SwapperMT.hpp
@@ -36,7 +36,7 @@ public:
     typedef std::complex<double> value_type;
     typedef SwapBuffers<value_type> swap_buffers_type;
 
-    static const size_t MaxGlobal;
+    static const uint64_t MaxGlobal;
 
     mpi::communicator world;
